requirements.txt
Flask
pydicom
Pillow